refactor: 클래스 패키지 분리
